test(engine): pin lifecycle dispatcher contracts — inline completion + pre-start cancellation
Retention review (#242 round 2):
- rename seam suspendDispatcher -> lifecycleDispatcher (drives ALL engine-owned
  lifecycleScope work: suspend invocations AND streaming); document contract
- keep the two strongest tests, drop the A1/A3 experimental controls (negative
  results preserved in PR history)
- remove repeat(100): deterministic tests arrange state, force transition,
  assert once
- bound ControlledDispatcher waits; assert close() terminated after join(5s)
- rename tests to claim completion/cancellation, not unprovable
  continuation-counting 'exactly once'
